backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==9.1.0
Pillow==10.1.0
python-dotenv==1.0.0